backend/requirements.txt
# Flask
fastapi
pandas
numpy
psycopg2-binary==2.9.1
jinja2
python-multipart